4B/week3/requirements.txt
# JKYog WhatsApp Bot - Requirements
# Complete dependencies for Render deployment

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Environment Variables
python-dotenv>=1.0.0

# OpenAI for Intent Classification
openai>=1.12.0
numpy>=1.24.0

# Database (Supabase PostgreSQL)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Twilio WhatsApp Integration
twilio>=8.10.0

# Google Maps API
googlemaps>=4.10.0

# Google Calendar API
google-auth>=2.0.0
google-api-python-client>=2.0.0

# Optional: Testing
# pytest>=7.4.0
# httpx>=0.25.0
